Several smaller bugfixes
- pd.Timedelta does not like to be initialized with `tz` argument,
failing with "Cannot pass a datetime or Timestamp with tzinfo with the
tz parameter. Use tz_convert instead."  My understanding is that since
all timezones are UTC throughout, it should be safe to drop this
everywhere.

- flph.fuzzylabels returns a list for me, and so I explicitly convert
it to an np.array before applying fancy indexing with `labels[pts] =
'GND'`.  Before, this would error with "list indices must be integers
or slices, not tuple".

- The existing glob pattern did not seem to work for the new way of
saving pickles into subdirectories of `outdir`.  Instead I'm now
applying to separate glob patterns that will pick up pkl files from
the `INDATA/` directory and its subdirectories.

- I'm creating the output directories for plots and data if they do
not exist already.

- The plotting functions mysteriously saw a `GN` label.  I'm really
not sure if this is the right thing to do here, but I simply assumed
that this was the same as `GND` and added another entry to the
colormap accordingly.  This may be really a symptom of a bug that
should be fixed in another place.

- In proc_fl, set `odir_np` also when do_save==False, because it's
used later on.